reestrutura o projeto colocando em módulos
