+++ .idea/modules.xml
@@ -2,8 +2,12 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
+      <module fileurl="file://$PROJECT_DIR$/Programacao_de_Tipos_de_Dados_Avancado/Exercicios/Exercicios.iml" filepath="$PROJECT_DIR$/Programacao_de_Tipos_de_Dados_Avancado/Exercicios/Exercicios.iml" />
+      <module fileurl="file://$PROJECT_DIR$/Programacao_de_Tipos_de_Dados_Avancado/Exercicios/Exercicios_02/Exercicios_02.iml" filepath="$PROJECT_DIR$/Programacao_de_Tipos_de_Dados_Avancado/Exercicios/Exercicios_02/Exercicios_02.iml" />
+      <module fileurl="file://$PROJECT_DIR$/Programacao_de_Tipos_de_Dados_Avancado/Exercicios/Exercicios_sobre_manipulacao_de_strings_e_vetores/Exercicios_sobre_manipulacao_de_strings_e_vetores.iml" filepath="$PROJECT_DIR$/Programacao_de_Tipos_de_Dados_Avancado/Exercicios/Exercicios_sobre_manipulacao_de_strings_e_vetores/Exercicios_sobre_manipulacao_de_strings_e_vetores.iml" />
       <module fileurl="file://$PROJECT_DIR$/Faculdade_Sistemas_de_Informacao.iml" filepath="$PROJECT_DIR$/Faculdade_Sistemas_de_Informacao.iml" />
       <module fileurl="file://$PROJECT_DIR$/Programacao_de_Tipos_de_Dados_Avancado/Programacao_de_Tipos_de_Dados_Avancado.iml" filepath="$PROJECT_DIR$/Programacao_de_Tipos_de_Dados_Avancado/Programacao_de_Tipos_de_Dados_Avancado.iml" />
+      <module fileurl="file://$PROJECT_DIR$/Programacao_de_Tipos_de_Dados_Avancado/Trabalhos/Trabalhos.iml" filepath="$PROJECT_DIR$/Programacao_de_Tipos_de_Dados_Avancado/Trabalhos/Trabalhos.iml" />
     </modules>
   </component>
 </project>
