+++ Programacao_de_Tipos_de_Dados_Avancado/Exercicios/Exercicios_sobre_manipulacao_de_strings_e_vetores/Exercicios_sobre_manipulacao_de_strings_e_vetores.iml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<module type="JAVA_MODULE" version="4">
+  <component name="NewModuleRootManager" inherit-compiler-output="true">
+    <exclude-output />
+    <content url="file://$MODULE_DIR$">
+      <sourceFolder url="file://$MODULE_DIR$/src" isTestSource="false" />
+    </content>
+    <orderEntry type="inheritedJdk" />
+    <orderEntry type="sourceFolder" forTests="false" />
+    <orderEntry type="module" module-name="Exercicios" />
+  </component>
+</module>
